+++ service/src/domain/adapter-message/matrix.message.adapter.module.ts
