+++ service/src/domain/adapter-message/matrix.message.adapter.ts
