+++ service/src/domain/adapter-room/dto/matrix.room.dto.message.request.community.ts
