+++ service/src/domain/adapter-room/matrix.room.adapter.module.ts
