+++ service/src/domain/adapter-room/type/matrix.room.options.ts
