+++ service/src/domain/adapter-room/type/matrix.room.ts
