+++ service/src/domain/adapter-user/matrix.user.adapter.module.ts
