+++ service/src/domain/agent/agent-factory/matrix.agent.factory.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
-import { MatrixMessageAdapterModule } from '../adapter-message/matrix.message.adapter.module';
-import { MatrixRoomAdapterModule } from '../adapter-room/matrix.room.adapter.module';
-import { MatrixUserAdapterModule } from '../adapter-user/matrix.user.adapter.module';
 import { MatrixAdminUserModule } from '@src/domain/matrix-admin/user/matrix.admin.user.module';
 import { MatrixAgentFactoryService } from './matrix.agent.factory.service';
+import { MatrixUserAdapterModule } from '../../adapter-user/matrix.user.adapter.module';
+import { MatrixRoomAdapterModule } from '../../adapter-room/matrix.room.adapter.module';
+import { MatrixMessageAdapterModule } from '../../adapter-message/matrix.message.adapter.module';
 
 @Module({
   imports: [
