+++ service/src/domain/agent/agent-pool/matrix.agent.pool.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { MatrixAgentPool } from '@src/domain/matrix/agent-pool/matrix.agent.pool';
+import { MatrixAgentPool } from '@src/domain/agent/agent-pool/matrix.agent.pool';
 import { MatrixAgentFactoryModule } from '../agent-factory/matrix.agent.factory.module';
 import { MatrixAdminUserModule } from '@src/domain/matrix-admin/user/matrix.admin.user.module';
