+++ service/src/domain/agent/agent/matrix.room.sliding.sync.window.manager.ts
