+++ service/src/domain/matrix-admin/user/matrix.admin.user.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { HttpModule } from '@nestjs/axios';
 import { MatrixCryptographyModule } from '@src/services/cryptography/matrix.cryptography.module';
-import { MatrixUserAdapterModule } from '@src/domain/matrix/adapter-user/matrix.user.adapter.module';
+import { MatrixUserAdapterModule } from '@src/domain/adapter-user/matrix.user.adapter.module';
 import { MatrixAdminUserService } from './matrix.admin.user.service';
 
 @Module({
