+++ service/src/services/communication-adapter/communication-adapter.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
-import { MatrixAgentPoolModule } from '@src/domain/matrix/agent-pool/matrix.agent.pool.module';
-import { MatrixRoomAdapterModule } from '@src/domain/matrix/adapter-room/matrix.room.adapter.module';
-import { MatrixUserAdapterModule } from '@src/domain/matrix/adapter-user/matrix.user.adapter.module';
+import { MatrixAgentPoolModule } from '@src/domain/agent/agent-pool/matrix.agent.pool.module';
+import { MatrixRoomAdapterModule } from '@src/domain/adapter-room/matrix.room.adapter.module';
+import { MatrixUserAdapterModule } from '@src/domain/adapter-user/matrix.user.adapter.module';
 import { CommunicationAdapter } from './communication.adapter';
 import { MatrixAdminUserElevatedModule } from '../../domain/matrix-admin/user-elevated/matrix.admin.user.elevated.module';
 import { MatrixAdminUserModule } from '../../domain/matrix-admin/user/matrix.admin.user.module';
